+++ internal/ceres/compressed_col_sparse_matrix_utils.h
@@ -51,20 +51,20 @@ namespace ceres::internal {
 // one non-zero entry corresponding to the top left entry of c_ij,
 // as that entry is used to detect the presence of a non-zero c_ij.
 CERES_NO_EXPORT void CompressedColumnScalarMatrixToBlockMatrix(
-    const int* scalar_rows,
-    const int* scalar_cols,
+    const int64_t* scalar_rows,
+    const int64_t* scalar_cols,
     const std::vector<Block>& row_blocks,
     const std::vector<Block>& col_blocks,
-    std::vector<int>* block_rows,
-    std::vector<int>* block_cols);
+    std::vector<int64_t>* block_rows,
+    std::vector<int64_t>* block_cols);
 
 // Given a set of blocks and a permutation of these blocks, compute
 // the corresponding "scalar" ordering, where the scalar ordering of
 // size sum(blocks).
 CERES_NO_EXPORT void BlockOrderingToScalarOrdering(
     const std::vector<Block>& blocks,
-    const std::vector<int>& block_ordering,
-    std::vector<int>* scalar_ordering);
+    const std::vector<int64_t>& block_ordering,
+    std::vector<int64_t>* scalar_ordering);
 
 // Solve the linear system
 //
